kg_project/requirements.txt
spacy
neo4j
groq
python-dotenv
fastapi
uvicorn
pymupdf
python-multipart
